+++ src/cards/Tutte/soccer-live-matches-editor.js
@@ -145,7 +145,7 @@ class SoccerLiveMatchesEditor extends LitElement {
 
     return html`
       <div class="card-config">
-        <h3>Sensor</h3>
+        <h3>${this._t("editor.sensor")}</h3>
         <div>
           <label class="field-label">${this._t('editor.entity')}</label>
           <select @change=${this._entityChanged}>
@@ -156,7 +156,7 @@ class SoccerLiveMatchesEditor extends LitElement {
           </select>
         </div>
 
-        <h3>Settings</h3>
+        <h3>${this._t("editor.settings")}</h3>
 
         <div>
           <label class="field-label">${this._t('editor.my_team')}</label>